requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
scipy>=1.7.0
tushare>=1.2.89
python-dotenv>=0.19.0
schedule>=1.2.0 